rename: Constants 폴더 구조 변경 - #33
